+++ src/components/QuizCreate/QuizSetForm/styles.tsx
@@ -1,18 +1,6 @@
 import styled from '@emotion/styled';
-import {
-  blackGray,
-  DarkGray,
-  lightGrayWhite,
-  pointColor,
-  small,
-  white,
-} from '@/styles/theme';
+import { DarkGray, lightGrayWhite, pointColor } from '@/styles/theme';
 
-export const FormContainer = styled.form`
-  width: 100%;
-  margin-top: 7rem;
-`;
-//* SetForm
 export const SetWrapper = styled.section`
   margin-bottom: 1rem;
   padding: 0.5rem 1.5rem;
@@ -73,14 +61,6 @@ export const SetTagInput = styled.input`
   }
 `;
 
-//* QuizListForm
-export const QuizListContainer = styled.section`
-  width: 100%;
-  display: flex;
-  flex-direction: column;
-  gap: 2rem;
-`;
-
 export const TextArea = styled.textarea`
   flex-basis: 100%;
   width: 100%;
@@ -91,37 +71,4 @@ export const TextArea = styled.textarea`
   border-radius: 0.5rem;
   font-family: 'Pretendard';
   resize: none;
-`;
-
-export const InsertQuizItem = styled.button`
-  height: 7.5rem;
-  display: flex;
-  justify-content: center;
-  align-items: flex-end;
-  padding-bottom: 1rem;
-
-  border: 3px dashed ${DarkGray};
-  border-bottom: none;
-  border-radius: 0.5rem 0.5rem 0 0;
-  background-color: ${white};
-
-  ${small};
-  color: ${blackGray};
-  cursor: pointer;
-`;
-
-export const SubmitButton = styled.button`
-  position: fixed;
-  right: 2rem;
-  bottom: 2rem;
-
-  width: 7.5rem;
-  height: 3rem;
-
-  border: 3px solid ${DarkGray};
-  border-radius: 0.5rem;
-  background-color: ${pointColor};
-  text-align: center;
-  font-family: 'MaplestoryOTFBold', sans-serif !important;
-  cursor: pointer;
 `;
